minor fixes, style fixed

+++ src/Authorization/EnhancedUserMapper.php
@@ -361,8 +361,7 @@ class EnhancedUserMapper extends UserMapper implements EnhancedUserMapperInterfa
 
         try {
             $pdos = $this->pdo->prepare(
-                'INSERT INTO user_role (user_id, role_id)
-            VALUES (:user_id, (SELECT role_id FROM role WHERE name = :role_name))'
+                'INSERT INTO user_role (user_id, role_id) VALUES (:user_id, (SELECT role_id FROM role WHERE name = :role_name))'
             );
 
             $pdos->bindParam(':user_id', $userId, PDO::PARAM_INT);
@@ -407,9 +406,7 @@ class EnhancedUserMapper extends UserMapper implements EnhancedUserMapperInterfa
         $userId = $user->getId();
 
         $pdos = $this->pdo->prepare(
-            'DELETE FROM user_role
-        WHERE role_id = (SELECT role_id FROM role WHERE name = :role_name) 
-        AND user_id = :user_id'
+            'DELETE FROM user_role WHERE role_id = (SELECT role_id FROM role WHERE name = :role_name) AND user_id = :user_id'
         );
 
         $pdos->bindParam(':role_name', $roleName, PDO::PARAM_STR);
